requirements.txt
flask
ultralytics
opencv-python
easyocr
numpy
pillow
